Fix stickes: line wrap mismatch between display/edit
The 'fake' textarea used to edit text in a sticky had a different
overflow-wrap style than the component that renders the text.

By forcing the display and edit components to use the same wrapping
strategy, the caret from the textarea and the rendered text should
remain in sync.

There is a chance there are more styles which affect the wrapping of
text that could still result in various mismatches, and even moreso
when we consider more browsers (I tested this in Chrome alone).